Add files via upload
My attempt at implementing a Neural Network, based off of some implementations I saw online and my own knowledge of the subject. All work done in python with the help of Pandas, Numpy and the random package. Unfortunately, some of the vectors were acting strangely and I am looking to discover the problem in the near future.